Renamed StandardProjectStructureGenerator to StandardCodeGenerator.
Creating Azzyztant and loading it in one step would not work, because
the project needs a refresh in order to make the compiler do its work,
but refreshing can't be done in MWE library. Thus we have split-off a
PrerequisiteGenerator. It is called first (generating Azzyztant), then
the plugin refreshes the projects, and finally the
StandardCodeGenerator is called.

Still, when loading a prerequisite class, we need to be prepared to
retry until loading succeeds, because the compiler may not have done
its work yet.